refactor(integrations): split booking/mod.rs into cohesive submodules
Behavior-preserving extraction of backend/crates/integrations/src/booking/mod.rs
(2786 lines) into focused submodules to reduce the churn hotspot's file size:

- error.rs    — BookingError
- models.rs   — domain types (credentials, property, reservation, guest, mappings)
- messages.rs — typed OTA request/response models + XML (de)serialisation
- client.rs   — the BookingClient HTTP client

mod.rs retains the module docs, OTA_NAMESPACE, the pre-existing ota_xml/oauth/
retry submodules, and the full test suite; all public items are re-exported so
booking::<Item> paths and the crate-level pub use in lib.rs resolve unchanged.
No functional change.
